dss: dss_device_get implementation bud
Currently only works with no criteria.
The function returns strings coming straight out of the postgres
result value, which needs to be freed later. It's good if we deal
with big json chucks but we might want to change that to copy if
we only keep small strings

Add a dss_res_free function to free memory allocated in dss_get
Add container_of and new string/enum conversion functions

Change-Id: Ied1c133b023e876d0675943e2b74ea895b8ab467