LRS: load device state and select a drive to perform write operation.
Retrieve the list of unlocked devices from the DSS and
query the system for their status, using a POC command.

Then select a drive according to a configurable policy
(first_fit or best_fit).

Tests behaves differently depending on the environment:
if tape drives are avaiblable and the /dev/changer is writable,
tests runs on LTFS filesystems.
Else (in particular if they are run by a user), tests use
directories of /tmp as a storage backend.

test of "GET" temporarily disabled until DSS implements extents support.

Change-Id: Id500ab65748479239900800270d1bfa7b6cdc3a9