test: avoid concurrent runs of 'make check'.
Added a script to create a test lock before running the test suite,
and manage a 5min timeout in case the test suite is aborted.
The lock file is removed when the test suite ends or after the
timeout.

Lock file is 777, which allows the lock file to be acquired by
a different user, even if it cannot delete it.

Change-Id: I7118824b1e37a4075699bed9323f5e40dccdfc65