ldm: add required types for new adapters.
This adds new types and accessors for new LDM interface
based on adapters to allow progressive review and landing.
This leaves previous functions to avoid breaking the build
and tests. Old code will be remove when the entire set of patches
for ldm adapters will be landed.
Conflicting function names have been temporarily prefixed with 'NEW_'.
This prefix will also be removed.

Change-Id: I234feaac53061b37836d896f67eb1420a585849b