tests: make 'test_store.sh' only test directories.
'test_store.sh' is a fast test running only on directories.

Tape access is now validated by 'acceptance.sh'.

The test is failing on directories because phobos takes
path as the directory id, but the test was expecting
id to be <host>:<path>.
Not sure it is the best direction to go, but at least this makes the
test and the code consistent.

Change-Id: I7a4c8ce107af90d237e868247d4ab498dd62f5c7