cli: properly add directories as (media, device)
A directory is a special object composed of a medium and a device.
Insert both from the CLI when adding new ones.

Fix small issues and remove LTFS-only checks in CLI/LRS/LDM layers
so that we can format a directory after adding it. This will not
issue an actual mkfs command but retrieve directory size only.

Change-Id: I0c40040afaf2da359d52ce305aad60ed9250b9d0
Phobos-Issue: PHO-101