lrs: fix media locking.
Media was not locked when the media was already loaded
in a drive, or if the filesystem was already mounted.
This resulted in errors or dangerous situations as
2 processes tried to write to the same media.

Also, in some error cases, the media was not unlocked.

Fixed these issues and slightly reorgonized some pieces of LRS code.

Added test of concurrent put in acceptance.sh.

Change-Id: I7d566f2c58ccbab89da250cc10605ebc7e20fa56