common: add retry loop macro.
This defines a macro to wrap retries of any function.
Retry strategy is defined by passing a custom retry function.

As a first use case, this implements a retry function for SCSI.
It perfoms short/delayed retries depending on the SCSI error code.

Added tests of retry loops in test_scsi.
Implementation in the LDM core will be done in a next patch.

Change-Id: I84816361a94f70183802d61ede5ec88e5535e730
Reviewed-on: https://cws-fleury.labs.ocre.cea.fr/gerrit/2006
Reviewed-by: Linter
Tested-by: Jenkins
Reviewed-by: Henri Doreau <henri.doreau@cea.fr>