ltfs: reduce available media space by 5%.
LTFS keeps some reserved space to flush its index but do not
reflect this in statfs().
The amount of reserved space is not clearly documented, but
it appears to be around 5% in some real cases.
So, reduce the free space returned by statfs() on LFTS by 5%.

This should address one of the aspects of RM#129.

Change-Id: I1f378add04f4d60bed9113640edfae9139dbac91
Reviewed-on: https://cws-fleury.labs.ocre.cea.fr/gerrit/2039
Reviewed-by: Linter
Tested-by: Jenkins
Reviewed-by: Henri Doreau <henri.doreau@cea.fr>