lib scsi: search a free slot to unload drive if none is specified
In some cases (e.g. after library power reset) the library doesn't provide
a source slot number for tapes loaded in drives.
In this case, we must search for an empty slot where the tape can be moved.

Completed test_scsi to check this case.
Also fixed it to properly handle drive serials.

Change-Id: I66bfcc820c5dd918f0689686bc3dd31d5fb12d25
Reviewed-on: https://cws-fleury.labs.ocre.cea.fr/gerrit/2093
Reviewed-by: Linter
Reviewed-by: Henri Doreau <henri.doreau@cea.fr>
Tested-by: Jenkins