lrs: don't leave a drive busy when its filesystem is full
When the filesystem served by a drive is detected as full,
don't leave the drive in busy state, as the drive could be used
to write another -non full- media instead.
This was problematic in the case all mounted FS were full:
the LRS could not select a drive to be unloaded as they were all tagged
busy.

Change-Id: Id72e5e6e4bfa3f2ecc0a6bdc19db99c395be94e9
Reviewed-on: https://cws-fleury.labs.ocre.cea.fr/gerrit/2638
Reviewed-by: Linter
Reviewed-by: Henri Doreau <henri.doreau@cea.fr>
Tested-by: Thomas Leibovici <thomas.leibovici@cea.fr>