ldm_lib_scsi: make tape move resilient to invalid slot address
In some cases, the source slot address of a tape cannot be used
as the move target, for example if it has been filled by another tape
in the meanwhile.

This fix prevents the lib_scsi LDM to move a tape to a full slot.
It also makes it resilient to such situation by retrying with another
free slot when the move failed with -EINVAL (likely invalid slot address).

Change-Id: Iae10212946771c3df889380062d3d2b68d27765e
Reviewed-on: https://cws-fleury.labs.ocre.cea.fr/gerrit/2651
Reviewed-by: Linter
Tested-by: Jenkins
Reviewed-by: Henri Doreau <henri.doreau@cea.fr>