pre-commit: remove broken commit hook
No modification can be done in the CLI w/o disabling this
check. It does not seem to be able to figure out its context
and produces hundreds of false-positives, hiding actual problems
and aborting legitimate commits.

Although it would be very good to have it, I rather explicitely
delete it for now until someone has time to fix it rather than
systematically have to disable it locally.

Change-Id: Iab62eefbacbfec3bcfdc62fc23e4354a9d8a42a9
Phobos-Issue: PHO-156
Reviewed-on: https://cws-fleury.labs.ocre.cea.fr/gerrit/2827
Reviewed-by: Linter
Tested-by: Jenkins
Reviewed-by: Jerome Andrieu <jerome.andrieu@cea.fr>
Reviewed-by: Henri Doreau <henri.doreau@cea.fr>