Add lib_scan and implement it for SCSI lib
This commit adds a new callback to struct lib_adapter called lib_scan.
This callback is designed for the sysadm to query various information
about the library. The callback is implemented for lib_adapter_scsi and
is made on what the scsi_mode_sense() function returns (usally use of library's
slots, states of drives and/or arms...).

Change-Id: Ic86fa9b27ffb631f9eb4efb5fe1872222c6a7159
Signed-off-by: Philippe DENIEL <philippe.deniel@cea.fr>
Reviewed-on: https://cws-fleury.labs.ocre.cea.fr/gerrit/3123
Reviewed-by: Linter
Tested-by: Jenkins
Tested-by: Jenkins s8open_nr <s8open_nr@ccc.ocre.cea.fr>
Reviewed-by: Patrice LUCAS <patrice.lucas@cea.fr>
Reviewed-by: Thomas Leibovici <thomas.leibovici@cea.fr>