dss: increase media id and device id size to NAME_MAX
media and device id size of 32 was too small given that they
are URIs including a hostname. It was also inconsistent with
the C code that handles a buffer of NAME_MAX.

This patch increases the DB field up to NAME_MAX.
Changed the C definition accordingly, to store (NAME_MAX + 1) bytes
including the null terminating byte.

Drop the union in media_id type and replace it by a single string
of NAME_MAX + 1 (defined as PHO_URI_MAX).

Phobos-Issue: PHO-390
Change-Id: Iaa08c2dfac1b8797a7988c9b28c9e2e72c995dbf
Signed-off-by: Thomas Leibovici <thomas.leibovici@cea.fr>
Reviewed-on: https://cws-fleury.labs.ocre.cea.fr/gerrit/3912
Reviewed-by: Linter
Reviewed-by: Florent Monjalet <florent.monjalet@cea.fr>
Tested-by: Jenkins s8open_nr <s8open_nr@ccc.ocre.cea.fr>