phobos.db: implement schema creation and deletion with tests
However, this code is not used yet to create and drop schemas, this
will come in a future commit.

This commit also allows to create and drop multiple versions of the
database schema (1.1 and 1.2).

The SQL scripts from pho_dss_helper have been put into separate files
and reformatted to be more readable and diff friendly.

Change-Id: I8009d44459777adbb65bf7a116d0023a3da6837f
Signed-off-by: Florent MONJALET <florent.monjalet@cea.fr>
Reviewed-on: https://cws-fleury.labs.ocre.cea.fr/gerrit/3983
Reviewed-by: Linter
Tested-by: Jenkins s8open_nr <s8open_nr@ccc.ocre.cea.fr>
Reviewed-by: Thomas Leibovici <thomas.leibovici@cea.fr>