lrs: force refreshing device ldm state for every IO
However, the DSS state is still cached. This is because raid1 calls
lrs_write_prepare twice and (indirectly, via intents) keeps pointers
to the lrs global device, hence this array cannot be reallocated.
Therefore, the choice has been made to refresh the "physical" state
of the drives, but not the DSS state (locking is checked on access).

Drives that are locked in the DSS are no longer filtered out, the
lock will be checked when phobos tries to acquire it.

This patch improves things, but there is still a possible race
condition if one phobos instance manages to change the media contained
in a drive between the moment where the other instance retrieves the
drive and media states and the moment it manages to lock it. There is
a need to check that the drive and media states are still satisfactory
after they have been locked.

The need to refresh the state of devices will allow to more correctly
retry IO on EAGAIN.

Change-Id: I0fb779eebe7b38886d7a8de94859ad225f07d2fb
Signed-off-by: Florent MONJALET <florent.monjalet@cea.fr>
Reviewed-on: https://cws-fleury.labs.ocre.cea.fr/gerrit/4082
Reviewed-by: Linter
Reviewed-by: Patrice LUCAS <patrice.lucas@cea.fr>
Tested-by: Jenkins s8open_nr <s8open_nr@ccc.ocre.cea.fr>
Reviewed-by: Thomas Leibovici <thomas.leibovici@cea.fr>