lrs: remove unused device parameter from lrs_select_media
lrs_select_media function doesn't use the device parameter. We
remove it to avoid misunderstanding about its behaviour.

For the moment, we choose a media first, without any device advise.
Then we pick a device accordingly to this chosen media.

This process is maybe not the final best one but this is the one we
currently use so it is better to clarify it.

Change-Id: Ie5249c496f1d2a112210763384ec3523f18f9fd6
Signed-off-by: Patrice LUCAS <patrice.lucas@cea.fr>
Reviewed-on: https://cws-fleury.labs.ocre.cea.fr/gerrit/4842
Reviewed-by: Linter
Tested-by: Jenkins s8open_nr <s8open_nr@ccc.ocre.cea.fr>
Reviewed-by: Florent Monjalet <florent.monjalet@cea.fr>
Reviewed-by: Thomas Leibovici <thomas.leibovici@cea.fr>