dss/lrs: lock owner name is now a parameter of `dss_*_{un,}lock`
Also, `dss_*_unlock` now check that the lock owner name was the one
passed as argument before unlocking. If the `lock_name` is `NULL`, the
existing lock owner name is not checked.

The lrs now contains its unique lock owner name. This prevents an lrs
from unlocking devices and media that it did not lock in the first
place.

A corollary of this change is that all public `lrs_*` function now
take a `struct lrs` as their first argument. As a side effect, `struct
intent` does not need to contain a dss reference anymore.

Change-Id: I7e2bd4e0c82072a6e350ae6e85d1ee20f0d41be3
Signed-off-by: Florent MONJALET <florent.monjalet@cea.fr>
Reviewed-on: https://cws-fleury.labs.ocre.cea.fr/gerrit/4877
Reviewed-by: Linter
Tested-by: Jenkins s8open_nr <s8open_nr@ccc.ocre.cea.fr>
Reviewed-by: Thomas Leibovici <thomas.leibovici@cea.fr>