Change store/layout/lrs interactions interfaces to use a protocol
This big refactoring changes the way store, layout and lrs interact,
switching from synchronous calls from layout to lrs to message passing
between the two (relayed by store). This refactoring is part of the
way to phobos 2.0 where the LRS will become a daemon.

This will allow:

- The LRS to implement smarter scheduling policy
- The layouts to implement more complex IO patterns
- The control flow not to be blocked when a layout attempts to reserve
  a media to read or write data.

Later on, the public API could be changed to expose asynchronicity
(poll/epoll-like API).

This version is a working draft:
- It compiles
- It passes tests (except for raid1, which has been temporarily
  removed)
- It removes temporarily raid1 (until it is ported in the very near
  future)
- It introduces performance regression (one flush per object vs
  one flush per mput)

Change-Id: I31a02caa6cd54d8b858b8ad6ce063ad7ff3ebece
Signed-off-by: Florent MONJALET <florent.monjalet@cea.fr>
Reviewed-on: https://cws-fleury.labs.ocre.cea.fr/gerrit/4967
Reviewed-by: Linter
Tested-by: Jenkins s8open_nr <s8open_nr@ccc.ocre.cea.fr>
Reviewed-by: Thomas Leibovici <thomas.leibovici@cea.fr>