checkpatch.pl: ignore ENOSYS kernel specific warning
In phobos, ENOSYS is used to signal that a layout plugin function
is not implemented, as errno(3) describes it as
"Function not implemented" (POSIX.1 description). However, checkpatch
enforces that it is only used for bad syscall number, which seems to
be a restriction concerning kernel code and not userland code.

Change-Id: I106f72171fff78640e52cccee252bad429509411
Signed-off-by: Florent MONJALET <florent.monjalet@cea.fr>
Reviewed-on: https://cws-fleury.labs.ocre.cea.fr/gerrit/5046
Reviewed-by: Linter
Tested-by: Jenkins s8open_nr <s8open_nr@ccc.ocre.cea.fr>
Reviewed-by: Patrice LUCAS <patrice.lucas@cea.fr>
Reviewed-by: Thomas Leibovici <thomas.leibovici@cea.fr>