acceptance.sh: replace fork & wait by xargs
Depending on bash configuration, it seems that there is a limit on
the number of background processes bash remembers. This can cause
a legitimate `wait $pid` to fail, as this PID has been since
forgotten by bash.

The workaround is to use xargs instead, which properly waits for its
children and has a non zero return code if any of them failed.

Change-Id: Idd02aff1e698cf03aecded07e40c10f14e0a3bfb
Signed-off-by: Florent MONJALET <florent.monjalet@cea.fr>
Reviewed-on: https://cws-fleury.labs.ocre.cea.fr/gerrit/5051
Reviewed-by: Linter
Tested-by: Jenkins s8open_nr <s8open_nr@ccc.ocre.cea.fr>
Reviewed-by: Patrice LUCAS <patrice.lucas@cea.fr>
Reviewed-by: Thomas Leibovici <thomas.leibovici@cea.fr>