lrs: remove misleading and redundant field in struct dev_descr
struct dev_descr used to have both a media_id and dss_media_info
field. dss_media_info->id contained the loaded media id, and
it had to be maintained consistent with media_id, which wasn't
the case and could cause some bugs when using a stale media_id
(as was the case in search_loaded_media). This patch removes the
misleading and redundant media_id field.

Change-Id: I462900883aea3e95484a357e4b359f6352a2bfc1
Signed-off-by: Florent MONJALET <florent.monjalet@cea.fr>
Reviewed-on: https://cws-fleury.labs.ocre.cea.fr/gerrit/5068
Reviewed-by: Linter
Tested-by: Jenkins s8open_nr <s8open_nr@ccc.ocre.cea.fr>
Reviewed-by: Patrice LUCAS <patrice.lucas@cea.fr>
Reviewed-by: Sebastien Gougeaud <sebastien.gougeaud@cea.fr>
Reviewed-by: Thomas Leibovici <thomas.leibovici@cea.fr>