store/layout/lrs: add serialization in the message protocol
Using protobuf-c, the requests and responses sent from/to the store
and the lrs are now serialized in a char buffer and deserialized
once received.

The 'proto' directory contains the protobuf files which are compiled
using the 'protoc-c' command at the start of phobos compilation.

The 'serializer' directory contains the protobuf structure helpers
for:
    - allocation/deallocation
    - packing/unpacking

Change-Id: I30979ebc073944677e0382323978e81762b57a68
Signed-off-by: Sebastien Gougeaud <sebastien.gougeaud@cea.fr>
Reviewed-on: https://cws-fleury.labs.ocre.cea.fr/gerrit/5096
Reviewed-by: Florent Monjalet <florent.monjalet@cea.fr>
Tested-by: Florent Monjalet <florent.monjalet@cea.fr>
Reviewed-by: Thomas Leibovici <thomas.leibovici@cea.fr>