store/layout: transfer requests/replies by unix socket
The requests and responses sent from/to the store and the lrs are
now transferred through a unix socket. The socket creation and
destruction are made during the {store,lrs}_{init,fini} calls.

The 'communication' directory contains the socket helpers.

The store module now triggers the lrs module with a lrs_process call to
let it retrieve store requests, process them and send back the
appropriate responses.

The lrs module is now designed as following:
- the scheduler part which mainly consists in the previous lrs module,
- the main part which mixes the scheduler with the communication and the
  serialization modules.
The lrs main part is the first step to its daemonization, with the
lrs_process function as its core function.

Change-Id: I2e621be80732a3a5ddd678a51737bc62ebceee0c
Signed-off-by: Sebastien Gougeaud <sebastien.gougeaud@cea.fr>
Reviewed-on: https://cws-fleury.labs.ocre.cea.fr/gerrit/5119
Reviewed-by: Linter
Tested-by: Jenkins s8open_nr <s8open_nr@ccc.ocre.cea.fr>
Reviewed-by: Patrice LUCAS <patrice.lucas@cea.fr>
Reviewed-by: Thomas Leibovici <thomas.leibovici@cea.fr>