admin: new module to launch admin phobos commands (format, device_add)
Because of LRS daemonization, a format or notify request needs to
be sent to the LRS via the transfer protocol. Also, as they are not
issued from user commands (format and device_add), they can not be
a part of phobos_store.

The phobos_admin module will be in charge to connect to the LRS socket
and send the format/notify request to it. Before the daemonization
is done, it will also intialize a temporary LRS to communicate with.

The notify request is for now used to notify the LRS that a device was
added to the DSS, and that it needs to load its state. This request can
then be used to inform the LRS about various state changes: like lock,
unlock, rm, etc.

Change-Id: I6c217cb701e99236e09676b53a62adff3a5691ad
Signed-off-by: Sebastien Gougeaud <sebastien.gougeaud@cea.fr>
Reviewed-on: https://cws-fleury.labs.ocre.cea.fr/gerrit/5157
Reviewed-by: Linter
Tested-by: Jenkins s8open_nr <s8open_nr@ccc.ocre.cea.fr>
Reviewed-by: Patrice LUCAS <patrice.lucas@cea.fr>
Reviewed-by: Thomas Leibovici <thomas.leibovici@cea.fr>