phobosd: LRS daemonization
The LRS module is now separated from the store/admin module and can be
reached through its unix socket. Also, it can not directly be called by
the CLI.

There are two ways of launching the daemon:
- as a daemon, with systemctl or executing phobosd;
- as an interactive executable (ie. attached to the terminal),
  with the -i option.

The tests are modified accordingly, using a helper script to
invoke/waive the daemon (test_launch_daemon.sh).

Change-Id: I1e0167f8ebcfbec2221f14a8143cd8b87f350937
Signed-off-by: Sebastien Gougeaud <sebastien.gougeaud@cea.fr>
Reviewed-on: https://cws-fleury.labs.ocre.cea.fr/gerrit/6337
Reviewed-by: Linter
Tested-by: Jenkins s8open_nr <s8open_nr@ccc.ocre.cea.fr>
Reviewed-by: Dominique Martinet <dominique.martinet@cea.fr>
Reviewed-by: Thomas Leibovici <thomas.leibovici@cea.fr>