lrs_sched: multi media split allocation
To allocate several media in split case and keeping safe the idea
of returning -EAGAIN if we have convenient media but currently
locked, we must track the context of a multi write allocation.

For example, if we need to book 2 media, one is already booked, we
are looking for the second one, we must not return -EAGAIN because
we think we should wait for the one we already booked for this
current allocation.

Change-Id: Ic7aa6dcc069ff5ace7253e10f62bf639bc1fa601
Signed-off-by: Patrice LUCAS <patrice.lucas@cea.fr>
Reviewed-on: https://cws-fleury.labs.ocre.cea.fr/gerrit/6392
Reviewed-by: Linter
Tested-by: Jenkins s8open_nr <s8open_nr@ccc.ocre.cea.fr>
Reviewed-by: Sebastien Gougeaud <sebastien.gougeaud@cea.fr>
Reviewed-by: Thomas Leibovici <thomas.leibovici@cea.fr>