tests: remove possibly-lost leaks from valgrind report
Possibly-lost leaks appear in valgrind reports when testing
acceptance.sh, generating millions of lines (each phobos command
execution leads to ~5k leaks). Those leaks are somewhat bound to
the python CLI, and are not clear enough to be resolve for now.

I made some research and there is a way to get better reports
and then try to resolve those leaks:
- get an environment where python is compiled in debug mode to
  let valgrind catch more information
$ ./configure --with-pydebug --without-pymalloc --with-valgrind

Change-Id: Ic861c1c8dcb73660aae400d32cc2b64908e0a9fd
Signed-off-by: Sebastien Gougeaud <sebastien.gougeaud@cea.fr>
Reviewed-on: https://cws-fleury.labs.ocre.cea.fr/gerrit/6403
Reviewed-by: Linter
Tested-by: Jenkins s8open_nr <s8open_nr@ccc.ocre.cea.fr>
Reviewed-by: Patrice LUCAS <patrice.lucas@cea.fr>
Reviewed-by: Thomas Leibovici <thomas.leibovici@cea.fr>