fix: assert encountered when adding a device without phobosd on
When inserting a device to the database, if the phobosd daemon is not
online, the admin module still tried to send a message to the daemon,
which leads to activate an assert (send a message but no active
communication socket).

This issue is now fixed by adding a boolean to the phobos admin handle,
indicating if the daemon is online or not. If not, then no message is
sent.

Change-Id: I2fe2d7b66550dd946b119f4f526652aa9fc845ed
Signed-off-by: Sebastien Gougeaud <sebastien.gougeaud@cea.fr>
Reviewed-on: https://cws-fleury.labs.ocre.cea.fr/gerrit/6475
Reviewed-by: Linter
Tested-by: Jenkins s8open_nr <s8open_nr@ccc.ocre.cea.fr>
Reviewed-by: Patrice LUCAS <patrice.lucas@cea.fr>
Reviewed-by: Thomas Leibovici <thomas.leibovici@cea.fr>