ldm_dev_scsi: Replace GSList with simple linked list implementation
GSList did trigger mysterious behaviors in glist_find_custom
(e.g. unexpected matching against NULL items).
The use of a Glib's slist is not needed here, as the slot allocation
is not performance-critical in this case.
A simple linked list which we control the implementation turns out
to be a better solution after all.

Change-Id: Id3daeb68acfb3a0fdea6beade0b420cf56e6c6c4
Signed-off-by: Thomas Leibovici <thomas.leibovici@cea.fr>
Reviewed-on: https://cws-fleury.labs.ocre.cea.fr/gerrit/6529
Reviewed-by: Linter
Tested-by: Jenkins s8open_nr <s8open_nr@ccc.ocre.cea.fr>
Reviewed-by: Sebastien Gougeaud <sebastien.gougeaud@cea.fr>
Reviewed-by: Quentin Bouget <quentin.bouget@cea.fr>
Reviewed-on: https://cws-fleury.labs.ocre.cea.fr/gerrit/6702