tests: add tests for phobosd bad requests reception
This test ensures that if phobosd receive bad requests, it returns an
error response and does not stop.

Some 'bad requests' (like write alloc for a null size, or read alloc
with a strange number of required devices) are not considered here as
they do not generate a processing error. Maybe change those behaviors in
future patchs.

Change-Id: I94bd0cdd28a93fde59e246a1cf1e9c4d0272723c
Signed-off-by: Sebastien Gougeaud <sebastien.gougeaud@cea.fr>
Reviewed-on: https://cws-fleury.labs.ocre.cea.fr/gerrit/6540
Reviewed-by: Linter
Tested-by: Jenkins s8open_nr <s8open_nr@ccc.ocre.cea.fr>
Reviewed-by: Quentin Bouget <quentin.bouget@cea.fr>
Reviewed-by: Thomas Leibovici <thomas.leibovici@cea.fr>