tests: wait phobosd termination before continuing tests
This modification ensures that the daemon is stopped before doing
anything else.

In the 'bad requests' test, we check if the daemon is running or not by
'ps' its PID. If we do not wait its closure, it may still exists,
cleaning its stuff before exiting. Thus, we consider it as existing,
which is in this case a wrong assumption.

Change-Id: I432b4835ab2109ef09fdc06efe43271fc58ff01b
Signed-off-by: Sebastien Gougeaud <sebastien.gougeaud@cea.fr>
Reviewed-on: https://cws-fleury.labs.ocre.cea.fr/gerrit/6539
Reviewed-by: Linter
Tested-by: Jenkins s8open_nr <s8open_nr@ccc.ocre.cea.fr>
Reviewed-by: Quentin Bouget <quentin.bouget@cea.fr>
Reviewed-by: Thomas Leibovici <thomas.leibovici@cea.fr>