refactor: unify medium and device identifiers
Because both media and devices are identified using a family and a
name, and to avoid code duplication, a resource is now identified by a
pho_id field. This data structure is a renamed version of the previous
media_id structure.

The name subfield stands for the resource primary key in the DB:
- the serial string for devices,
- the ident string for media.

This pho_id field and the model field are now part of a Resource data
structure which contains information that are part of both media and
devices.

This patch unifies the identifiers for the phobos source code, but not
the protobuf protocol. This one will be done in the next patch.

Change-Id: I2c7de7fa935986b91943299cb82b97778528ac0c
Signed-off-by: Sebastien Gougeaud <sebastien.gougeaud@cea.fr>
Reviewed-on: https://cws-fleury.labs.ocre.cea.fr/gerrit/6541
Reviewed-by: Linter
Tested-by: Jenkins s8open_nr <s8open_nr@ccc.ocre.cea.fr>
Reviewed-by: Quentin Bouget <quentin.bouget@cea.fr>
Reviewed-by: Thomas Leibovici <thomas.leibovici@cea.fr>