scheduler: fix sched_init() behavior in case of error
Before this patch, if an error happened during the initialization of a
scheduler, it may lead to a memory leak: dss data structure is not freed
if the allocation of lock_owner failed.

This behavior is prevented by reordering the initialization instructions
and freeing allocated data if an error occurs.

Change-Id: I31e9b7417815b230a7f33671b438ce44a073117c
Signed-off-by: Sebastien Gougeaud <sebastien.gougeaud@cea.fr>
Reviewed-on: https://cws-fleury.labs.ocre.cea.fr/gerrit/6651
Reviewed-by: Linter
Tested-by: Jenkins s8open_nr <s8open_nr@ccc.ocre.cea.fr>
Reviewed-by: Quentin Bouget <quentin.bouget@cea.fr>
Reviewed-by: Thomas Leibovici <thomas.leibovici@cea.fr>