tests: reorder database setup and daemon invocation
Directly inserting resources in the database does not notify the daemon
that new resources are available. The daemon will therefore retrieve
device information once it started processing a request.

It will lead to an issue in a future patch when we check that the
requested number of resources is available before processing a request.
For that reason, and because the database inserts do not involve the
daemon, its invocation is done once the directory setup is done.

Change-Id: I095859e7965b496b63657466c3d1b0361dd34dbf
Signed-off-by: Sebastien Gougeaud <sebastien.gougeaud@cea.fr>
Reviewed-on: https://cws-fleury.labs.ocre.cea.fr/gerrit/6655
Reviewed-by: Linter
Reviewed-by: Quentin Bouget <quentin.bouget@cea.fr>
Tested-by: Jenkins s8open_nr <s8open_nr@ccc.ocre.cea.fr>
Reviewed-by: Thomas Leibovici <thomas.leibovici@cea.fr>