refactor: aggregate XFer PUT parameters in a new structure
To avoid the xfer_desc structure pollution with an increasing set
of parameters (with the incoming addition of family and layout options),
the PUT-only parameters are now aggregated in a pho_xfer_put_params
data structure.

This new data structure is part of a union in the xfer_desc structure to
allow the future utilization of GET/GETMD sets of parameters.

Change-Id: I0063611df3c87bf53c8c1d783741112a6a2faac0
Signed-off-by: Sebastien Gougeaud <sebastien.gougeaud@cea.fr>
Reviewed-on: https://cws-fleury.labs.ocre.cea.fr/gerrit/6663
Reviewed-by: Linter
Reviewed-by: Patrice LUCAS <patrice.lucas@cea.fr>
Tested-by: Jenkins s8open_nr <s8open_nr@ccc.ocre.cea.fr>
Reviewed-by: Thomas Leibovici <thomas.leibovici@cea.fr>