tests: requests processing depending on resource availability
This test ensures that a request succeeds or fails according to the
resources that are available or not (non existent, unformatted,
locked).

As tape and directory families do not require the same resources (drive
or dummy device), they are both tested. It is possible that other
future families also need to be tested here, as their scheduler
implementation may differ.

TODO: For now, drive tests are commented, until the patch which fixes
locked device selection is on.

Change-Id: I2838b4c41d04d18e7563e9e93693d589f8b60b9b
Signed-off-by: Sebastien Gougeaud <sebastien.gougeaud@cea.fr>
Reviewed-on: https://cws-fleury.labs.ocre.cea.fr/gerrit/6682
Reviewed-by: Linter
Tested-by: Jenkins s8open_nr <s8open_nr@ccc.ocre.cea.fr>
Reviewed-by: Patrice LUCAS <patrice.lucas@cea.fr>
Reviewed-by: Thomas Leibovici <thomas.leibovici@cea.fr>