cli: 'list -o' prints following user order
Except for xml format (where the dictionary insertion order is
preserved in python3.6 and above), the output of 'list -o' commands
is now ordered following the attributes order given by the user.

Because csv format order is not an alphabetical one anymore,
acceptance.sh 'list -o' commands are modified if followed by a grep,
to only output the requested attribute.

Change-Id: I9b667d969fae955137d83cb1d22c7231f5c1caa1
Signed-off-by: Sebastien Gougeaud <sebastien.gougeaud@cea.fr>
Reviewed-on: https://cws-fleury.labs.ocre.cea.fr/gerrit/6708
Reviewed-by: Linter
Reviewed-by: Patrice LUCAS <patrice.lucas@cea.fr>
Tested-by: Jenkins s8open_nr <s8open_nr@ccc.ocre.cea.fr>
Reviewed-by: Thomas Leibovici <thomas.leibovici@cea.fr>