cli: add daemon notify when adding a dir (dummy device)
A faulty behavior could happen when formatting directories if the daemon
did not add their corresponding dummy devices. Now, if a directory is
added, the daemon is notified.

Also fix missing layout in test_resource_availability.sh. This cannot be
done in another patch, as the tests need both modifications to succeed.

Change-Id: I8ed551389fb548a6ceb523cb6dce68d4b84ea47b
Signed-off-by: Sebastien Gougeaud <sebastien.gougeaud@cea.fr>
Reviewed-on: https://gerrit.ccc.ocre.cea.fr/gerrit/6788
Reviewed-by: Linter
Tested-by: Jenkins s8open_nr <s8open_nr@ccc.ocre.cea.fr>
Reviewed-by: Quentin Bouget <quentin.bouget@cea.fr>
Reviewed-by: Thomas Leibovici <thomas.leibovici@cea.fr>