scheduler: fix deadlock during device selection
During the search of a device to free, if no device were found, the
scheduler checked if a compatible device exists. In this case, it returned
EBUSY, waiting for this device to be available. If this device was locked
for the current operation (write on RAID1 for instance), the scheduler
indefinitely waited.

Now, it does not consider the selected devices as compatibles ones
anymore, and returns ENODEV.

Change-Id: I81cfa184bfec5a584b86174864d9a4437a5ab043
Signed-off-by: Sebastien Gougeaud <sebastien.gougeaud@cea.fr>
Reviewed-on: https://gerrit.ccc.ocre.cea.fr/gerrit/c/phobos/+/6887
Tested-by: Jenkins s8open_nr <s8open_nr@ccc.ocre.cea.fr>
Reviewed-by: Thomas Leibovici <thomas.leibovici@cea.fr>
Reviewed-by: Quentin Bouget <quentin.bouget@cea.fr>
Reviewed-by: Patrice LUCAS <patrice.lucas@cea.fr>