cli: fix bad directory addition
Previously, if a non-existing directory was added, it was still added
to the media table in the database.

Now, a non-existing device is not added. And if a 'phobos dir add'
command includes good and bad directory names, the good ones are added
and error messages indicate which directories can not be added.

Change-Id: I28a974965ece3c5fb9fb493143e5ca5240e1d1e3
Signed-off-by: Sebastien Gougeaud <sebastien.gougeaud@cea.fr>
Reviewed-on: https://gerrit.ccc.ocre.cea.fr/gerrit/c/phobos/+/7287
Tested-by: Jenkins s8open_nr <s8open_nr@ccc.ocre.cea.fr>
Reviewed-by: Patrice LUCAS <patrice.lucas@cea.fr>
Reviewed-by: Thomas Leibovici <thomas.leibovici@cea.fr>