fix: configuration file loading in PUT requests
During the initialization of XFer data structures for PUT requests, if
the family or the layout were not given by the user, the default values
were retrieved from the configuration module. This module will then
retrieve the value from the environment, then the configuration file and
finally the source code. Because the configuration file was not loaded
yet, its default values were not taken into account.

Now, the configuration module is initialized before retrieving the
default values.

Change-Id: I6012f4aca40f3bc3095a29835efccf5f4506971a
Signed-off-by: Sebastien Gougeaud <sebastien.gougeaud@cea.fr>
Reviewed-on: https://gerrit.ccc.ocre.cea.fr/gerrit/c/phobos/+/7732
Tested-by: Jenkins s8open_nr <s8open_nr@ccc.ocre.cea.fr>
Reviewed-by: Quentin Bouget <quentin.bouget@cea.fr>
Reviewed-by: Thomas Leibovici <thomas.leibovici@cea.fr>