python3: change division operator to floor division
In python2, the division operator outputs a value of the type given by
the inputs (so int/int gives an int). In python3, the output value is a
float one, which provokes an error with the logging package because its
logging level needs to be an integer. Instead, we now use the floor
division operator (//).

Change-Id: I66c6286421de3ef3a12e4afc745900cf2e9feb9e
Signed-off-by: Sebastien Gougeaud <sebastien.gougeaud@cea.fr>
Reviewed-on: https://gerrit.ccc.ocre.cea.fr/gerrit/c/phobos/+/7767
Reviewed-by: Quentin Bouget <quentin.bouget@cea.fr>
Reviewed-by: Patrice LUCAS <patrice.lucas@cea.fr>
Tested-by: Jenkins s8open_nr <s8open_nr@ccc.ocre.cea.fr>