Add alias functionality
An alias is a name for a combination of the put parameters family, layout and tags.
This can be created in the config file and then added to the put parameters and with the CLI option -a or --alias.
If set it will replace the default parameters from the alias if not otherwise specified (ie. if and alias and a family is set the alias will only replace tags and layout in the parameters).

Change-Id: I2c54466b0f0f19af55d75d5ef527a412e93fe60e