Add alias functionality
An alias is a name for a combination of the put parameters: family, layout and tags.
This can be created in the config file and then added to the put parameters with the CLI option -a or --alias.
Direct command line family and layout options supersede alias parameters (ie. if an alias and a family are set, the alias family if it exists will be ignored). Direct command line tags and alias tags are joined.

Change-Id: I2c54466b0f0f19af55d75d5ef527a412e93fe60e
Signed-off-by: Sophie Wenzel-Teuber <sophie.wenzel-teuber@ichec.ie>