Update patch set 1
Patch Set 1: Code-Review+1

Thanks for this patch. You can rebase the alias patch on top this one.

(More globally, I was reflecting if the "INVAL" value is a good name ... maybe UNSET is more clear and maybe the 0 value could be better for an API where by default we set a parameter ... But this modification, is beyond this patch.)

Patch-set: 1
Reviewer: Gerrit User 1008040 <1008040@d5d70762-12d0-45a1-890d-524b12d3f735>
Label: Code-Review=+1