DB: 1.91, towards deletion and versionning
We add version and uuid fields into the object table and the
extent table. The primary key of the object table remains oid
whereas the primary key of the extent table moves from oid to
(uuid, version). We create the new deprecated_object table.

This patch contains only the update of the database structure.
UUID, version fields and deprecated_objects are still ignored
and unused by the phobos code but the database is now ready for
the next step.

Change-Id: Ib0877ae6841f4f9962810b309633d18da9edefad
Signed-off-by: Patrice LUCAS <patrice.lucas@cea.fr>
Reviewed-on: https://gerrit.ccc.ocre.cea.fr/gerrit/c/phobos/+/7968
Reviewed-by: Sebastien Gougeaud <sebastien.gougeaud@cea.fr>
Reviewed-by: Quentin Bouget <quentin.bouget@cea.fr>
Tested-by: Jenkins s8open_nr <s8open_nr@ccc.ocre.cea.fr>