store: add uuid and version fields to the XFer structure
I/O operations need to know object UUID and version number to generate a
unique extent path, and the only data structure used by the I/O adapter
is an XFer structure.

We also remove the undel parameters structure of the XFer, as the field
it contains is redundant with the new xd_objuuid field.

Change-Id: I2d0d0aeb18419131cb7dddf5286e5de9f062bd92
Signed-off-by: Sebastien Gougeaud <sebastien.gougeaud@cea.fr>
Reviewed-on: https://gerrit.ccc.ocre.cea.fr/gerrit/c/phobos/+/8226
Tested-by: Jenkins s8open_nr <s8open_nr@ccc.ocre.cea.fr>
Reviewed-by: Thomas Leibovici <thomas.leibovici@cea.fr>