io: change io extent path generation
The previous generation function took an oid as a primary key, which
leads to issues when putting a new version of an object. The new process
now considers the triplet <version, extent_tag, uuid> as the primary
key.

As the key must now contains '.' to delimit its fields, some tests are
removed.

Change-Id: I150f735cf14b53133293e6b6e14a98929869163e
Signed-off-by: Sebastien Gougeaud <sebastien.gougeaud@cea.fr>
Reviewed-on: https://gerrit.ccc.ocre.cea.fr/gerrit/c/phobos/+/8227
Tested-by: Jenkins s8open_nr <s8open_nr@ccc.ocre.cea.fr>
Reviewed-by: Thomas Leibovici <thomas.leibovici@cea.fr>