cli: fix mput file parser
The original parsing for the mput command was matching a
file line with a regexp so that if the file name or oid
included spaces (escaped or not), the information would
slip into metadata, creating errors on files and oids.
We changed the parser to make it POSIX compliant, detecting
for spaces, and then checking there
are only 3 "valid" items written on each line.

Change-Id: Id65bfa4fc9ba7024341fdc7f174fe2d0e5a3311a
Signed-off-by: Yoann Valeri <yoann.valeri@cea.fr>
Reviewed-on: https://gerrit.ccc.ocre.cea.fr/gerrit/c/phobos/+/8289
Reviewed-by: Patrice LUCAS <patrice.lucas@cea.fr>
Reviewed-by: Sebastien Gougeaud <sebastien.gougeaud@cea.fr>
Tested-by: Jenkins s8open_nr <s8open_nr@ccc.ocre.cea.fr>