dss: use new lock API for lock/unlock on device/media
This patch aims to be the basis of a chain of
modification to the tests using the concurrency
locks. It currently uses both the new and old
locks, until every test is modified to only
use the new table.
Also change PhobosDSSTest as we decided that
unlocking twice shouldn't be possible, which
was allowed before.

Change-Id: I0d3c5b53869034a9eedbe355ee31d0297ef73011
Signed-off-by: Yoann Valeri <yoann.valeri@cea.fr>
Reviewed-on: https://gerrit.ccc.ocre.cea.fr/gerrit/c/phobos/+/8396
Tested-by: Jenkins s8open_nr <s8open_nr@ccc.ocre.cea.fr>
Reviewed-by: Patrice LUCAS <patrice.lucas@cea.fr>
Reviewed-by: Sebastien Gougeaud <sebastien.gougeaud@cea.fr>
Reviewed-by: Guillaume Courrier <guillaume.courrier@cea.fr>