dss, lrs and cli: consolidate get using new lock API
Modify the CLI get command to properly use the fields of
the new lock table.
Modify 'phobos_generic_get' to check for concurrency
locks in the table and update the result.
Modify the LRS to properly use the new fields and
structures.
Modify test_daemon.sh to use the new table and locks.

Change-Id: Ic9b1a083f92e356aa3601738286c02a8b9e26518
Signed-off-by: Yoann Valeri <yoann.valeri@cea.fr>
Reviewed-on: https://gerrit.ccc.ocre.cea.fr/gerrit/c/phobos/+/8412
Tested-by: Jenkins s8open_nr <s8open_nr@ccc.ocre.cea.fr>
Reviewed-by: Guillaume Courrier <guillaume.courrier@cea.fr>
Reviewed-by: Patrice LUCAS <patrice.lucas@cea.fr>