cli: fix logger cleaning
Each python instance of PhobosActionContext is modifying the
phobos_log_callback static global variable of the dss.c library
with a value wich is only valid until this PhobosActionContext
instance exists. So, when we go out of the PhobosActionContext
it is mandatory to clean the C phobos_log_callback value.

Change-Id: I424e279f5501b0c5f14dcd45cbf2ffef8338f377
Signed-off-by: Patrice LUCAS <patrice.lucas@cea.fr>
Reviewed-on: https://gerrit.ccc.ocre.cea.fr/gerrit/c/phobos/+/8429
Reviewed-by: Yoann VALERI <yoann.valeri@cea.fr>
Reviewed-by: Sebastien Gougeaud <sebastien.gougeaud@cea.fr>
Tested-by: Jenkins s8open_nr <s8open_nr@ccc.ocre.cea.fr>