fix: improper use of dss_{un}lock in object_{un}delete
The previous version of this code was working because
since we were giving it a DSS_OBJECT to dss_{un}lock,
it would fetch the oid field of a supposed object_info,
which is the first one.
Therefore, since we were casting a char* as an
object_info*, the parser would work properly and find
the name associated with the object we want to {un}lock.
This patch modifies this behaviour to make it consistent
with the dss_{un}lock API.

Change-Id: I6712dfa9d4eb2026ee493a5677953c51a4ffb148
Signed-off-by: Yoann Valeri <yoann.valeri@cea.fr>
Reviewed-on: https://gerrit.ccc.ocre.cea.fr/gerrit/c/phobos/+/8475
Tested-by: Jenkins s8open_nr <s8open_nr@ccc.ocre.cea.fr>
Reviewed-by: Patrice LUCAS <patrice.lucas@cea.fr>
Reviewed-by: Sebastien Gougeaud <sebastien.gougeaud@cea.fr>
Reviewed-by: Guillaume Courrier <guillaume.courrier@cea.fr>